backend/requirements.txt
python-dotenv
langchain-text-splitters
langchain-community
langchain-google-genai
unstructured
unstructured-client
unstructured-inference
unstructured[pdf]
faiss-gpu
sentence-transformers
nltk
transformers
docx2txt
python-docx
openpyxl
ocrmypdf
poppler-utils
python-multipart
uuid
faiss-cpu
fastapi
uvicorn
pydantic
requests
beautifulsoup4
selenium
webdriver-manager
langchain
pypdf
youtube-transcript-api
SpeechRecognition
pydub
moviepy
langdetect
googletrans